Update to LFS v1.7.2 tag (see below for details) I wanted to update LFS to pull in a particular fix they made upstream: https://github.com/littlefs-project/littlefs/commit/97d8d5e96a7781596708664f18f2ea6c3a179330
Since I was doing the update anyways (to this old/stable lib) I updated to
the 'highest' LFS version which is still backwards compatible to the
v1.6.1 version which had been used originally to create the adafruit fork.

The other fixes in 1.7.2/1.7.0 (they are identical wrt. lfs.c) IMO are
unlikely to be critical for adafruit but see the full list here:

https://github.com/littlefs-project/littlefs/releases/tag/v1.7.0